Backport patches from drop 48.1

+++ maint/jenkins/scripts/rpm-testing.sh
@@ -20,8 +20,7 @@ force_am=noam
 
 MPI_DIR=""
 ze_dir=""
-GENGBIN_NEO=/home/gengbinz/drivers.gpu.compute.runtime/workspace-09-10-2021
-if [ "$flavor" == "ats" ]; then
+if [ "$flavor" == "gpu" ]; then
     ze_dir="/usr"
 fi
 
@@ -35,7 +34,7 @@ if [ "$flavor" != "regular" ]; then
     flavor_string="-${flavor}"
 fi
 
-if [ "$flavor" == "ats" ]; then
+if [ "$flavor" == "gpu" ]; then
     config_opts="--with-ze=${ze_dir}"
 fi
 
@@ -55,7 +54,7 @@ export MPITEST_TIMEOUT_MULTIPLIER=2.0
 
 JENKINS_DIR="$WORKSPACE/maint/jenkins"
 BUILD_SCRIPT_DIR="$JENKINS_DIR/scripts"
-if [ "${flavor}" != "ats" ]; then
+if [ "${flavor}" != "gpu" ]; then
     OFI_DIR="/opt/intel/csr/ofi/${provider}-dynamic"
 else
     if [ "${provider}" == "psm3" ]; then
@@ -94,7 +93,7 @@ count=1
 rpm -qpR ${WORKSPACE}/$RPM.rpm
 
 set -e
-if [ "$flavor" == "ats" ]; then
+if [ "$flavor" == "gpu" ]; then
     set +e
     rpm --initdb --dbpath /tmp/rpmdb
     rpm --dbpath /tmp/rpmdb --nodeps -ihv --prefix /tmp/inst ${WORKSPACE}/$RPM.rpm
@@ -104,7 +103,7 @@ else
 fi
 
 if [ "$nodes" -gt "1" ]; then
-    if [ "$flavor" == "ats" ]; then
+    if [ "$flavor" == "gpu" ]; then
         srun -N 1 -n 1 -r 1 $BUILD_SCRIPT_DIR/install_drop_rpm.sh 0 ${WORKSPACE}/$RPM.rpm
     else
     	srun -N 1 -n 1 -r 1 sudo rpm -Uvh --force ${WORKSPACE}/$RPM.rpm --nodeps
@@ -154,7 +153,7 @@ fi
 # Make sure module alias is setup
 . /opt/ohpc/admin/lmod/lmod/init/bash >/dev/null
 # check env by using module
-if [ "$flavor" == "ats" ]; then
+if [ "$flavor" == "gpu" ]; then
     module use /tmp/inst/modulefiles/
     MPI_DIR="/tmp/inst/${MPI}"
 else
@@ -193,7 +192,7 @@ fi
 
 
 # This is used on the anfedclx8 machine
-if [ "$flavor" == "ats" ]; then
+if [ "$flavor" == "gpu" ]; then
     export LD_LIBRARY_PATH=/usr/lib64:$LD_LIBRARY_PATH
 else
     export LD_LIBRARY_PATH=/usr/lib64:/opt/dg1/clan-spir-1.1/lib:$LD_LIBRARY_PATH
@@ -201,11 +200,6 @@ fi
 
 export LD_LIBRARY_PATH=${OFI_DIR}/lib:/opt/intel/csr/lib:/opt/intel/csr/lib64:$LD_LIBRARY_PATH
 export LD_LIBRARY_PATH=/usr/local/lib:$LD_LIBRARY_PATH
-
-# Make sure the patched NEO is in front in LD_LIBRARY_PATH if on ats cluster
-if [ -d "$GENGBIN_NEO" ]; then
-    export LD_LIBRARY_PATH=/home/gengbinz/drivers.gpu.compute.runtime/workspace-09-10-2021/neo/build/bin:/home/gengbinz/drivers.gpu.compute.runtime/workspace-09-10-2021/neo/build/lib:/home/gengbinz/drivers.gpu.compute.runtime/workspace-09-10-2021/igc/lib:/home/gengbinz/drivers.gpu.compute.runtime/workspace-09-10-2021/gmmlib/lib:$LD_LIBRARY_PATH
-fi
 
 export PATH=/opt/intel/csr/bin:$PATH
 
@@ -240,11 +234,19 @@ fi
 
 echo "checking automake..."
 echo $PATH
-which automake
-./autogen.sh --without-ucx
+
+AUTOTOOLS_DIR="/home/sys_csr1/software/autotools/bin"
+export PATH=$AUTOTOOLS_DIR:$PATH
+
+if [ ! -d /home/sys_csr1/software ] ; then
+    exit 1;
+fi
+
+which ${AUTOTOOLS_DIR}/automake
+./autogen.sh --with-autotools=${AUTOTOOLS_DIR} --without-ucx
 #from test dir
 cd test/mpi
-./autogen.sh
+./autogen.sh --with-autotools=${AUTOTOOLS_DIR}
 
 if [ "$testgpu" == "0" ]; then
     find . -name testlist.gpu -exec rm '{}' \; -and -exec touch '{}' \;
@@ -264,6 +266,7 @@ fi
 
 ldd ${MPI_DIR}/lib/libmpich.so
 
+echo "rpm-testing: slurm allocated node is: "`srun -N 1 -n 1 hostname`
 echo "rpm-testing: test/mpi/configure will run on hostname: ${HOSTNAME}"
 
 ./configure --with-mpi=${MPI_DIR} --disable-perftest --disable-ft-tests ${error_checking} ${config_opts} \
@@ -288,7 +291,7 @@ python3 ${WORKSPACE}/set_xfail.py -j validation -c ${compiler} -o ${configs} -s 
     -m "${provider_string}" -a ${force_am} -f ${xfail_file} -p ${pmix}
 
 # For the GPU clusters, we need to additionally xfail all of the known problems with GPUs
-if [ "$flavor" == "ats" ]; then
+if [ "$flavor" == "gpu" ]; then
     python3 ${WORKSPACE}/set_xfail.py -j per-commit-gpu -c ${compiler} -o ${configs} -s ${direct} \
         -m "${provider_string}" -a ${force_am} -f ${xfail_file} -p ${pmix}
 fi
@@ -306,7 +309,7 @@ if [ $? != 0 ]; then
 fi
 cp summary.junit.xml ${WORKSPACE}
 
-if [ "$flavor" == "ats" ]; then
+if [ "$flavor" == "gpu" ]; then
     rm -rf /tmp/inst
     rm -rf /tmp/rpmdb
 else
@@ -314,7 +317,7 @@ else
 fi
 
 if [ "$nodes" -gt "1" ]; then
-   if [ "$flavor" == "ats" ]; then
+   if [ "$flavor" == "gpu" ]; then
        srun -N 1 -n 1 -e 1 $BUILD_SCRIPT_DIR/install_drop_rpm.sh 1 ${WORKSPACE}/$RPM.rpm
    else
        srun -N 1 -n 1 -r 1 sudo rpm -e $RPM
